+++ apps/sif-ettersending/src/app/application/ApplicationRoutes.tsx
@@ -1,22 +1,22 @@
 import * as React from 'react';
-import { Redirect, Route, Switch, useHistory } from 'react-router-dom';
+import { Navigate, Route, Routes, useNavigate } from 'react-router-dom';
 import { useAmplitudeInstance } from '@navikt/sif-common-amplitude/lib';
 import { useFormikContext } from 'formik';
 import ConfirmationPage from '../components/pages/confirmation-page/ConfirmationPage';
 import GeneralErrorPage from '../components/pages/general-error-page/GeneralErrorPage';
 import WelcomingPage from '../components/pages/welcoming-page/WelcomingPage';
-import { getRouteConfig, getRouteUrl } from '../config/routeConfig';
-import { StepID, getFirstStep } from '../config/stepConfig';
+import { getRouteConfig, APPLICATION_SENDT_PAGE, ERROR_PAGE } from '../config/routeConfig';
+import { getFirstStep, StepID } from '../config/stepConfig';
 import { ApplicationTypeContext } from '../context/ApplicationTypeContext';
 import { ApplicationFormData } from '../types/ApplicationFormData';
 import { ApplicationType } from '../types/ApplicationType';
 import { getSkjemanavn } from '../types/skjemanavn';
-import { navigateTo } from '../utils/navigationUtils';
-import { getApplicationRoute, getNextStepRoute, isAvailable } from '../utils/routeUtils';
+import { getNextStepRoute, isStepAvailable } from '../utils/routeUtils';
 import BeskrivelseStep from './beskrivelse-step/BeskrivelseStep';
 import DokumenterStep from './dokumenter-step/DokumenterStep';
 import OppsummeringStep from './oppsummering-step/OppsummeringStep';
 import ValgOmsTypeStep from './valgOmsType-step/ValgOmsTypeStep';
+import { relocatoToKvitteringPage } from '../utils/navigationUtils';
 
 export interface KvitteringInfo {
     søkernavn: string;
@@ -26,97 +26,85 @@ const ApplicationRoutes = () => {
     const { values } = useFormikContext<ApplicationFormData>();
     const { søknadstype } = React.useContext(ApplicationTypeContext);
 
-    const history = useHistory() as any;
+    const navigate = useNavigate();
     const { logSoknadStartet } = useAmplitudeInstance();
 
     if (!søknadstype) {
-        return <Route path={getRouteConfig(ApplicationType.ukjent).ERROR_PAGE_ROUTE} component={GeneralErrorPage} />;
+        return <Route path={getRouteConfig(ApplicationType.ukjent).ERROR_PAGE_ROUTE} element={<GeneralErrorPage />} />;
     }
-    const routeConfig = getRouteConfig(søknadstype);
 
     const navigateToNextStep = (stepId: StepID) => {
         setTimeout(() => {
             const nextStepRoute = getNextStepRoute(søknadstype, stepId);
             if (nextStepRoute) {
-                navigateTo(nextStepRoute, history);
+                navigate(nextStepRoute);
             }
         });
     };
 
     const startSoknad = async () => {
         await logSoknadStartet(getSkjemanavn(søknadstype));
+        const firstStep = getFirstStep(søknadstype);
         setTimeout(() => {
-            navigateTo(`${routeConfig.APPLICATION_ROUTE_PREFIX}/${getFirstStep(søknadstype)}`, history);
+            navigate(firstStep);
         });
     };
 
     return (
-        <Switch>
-            <Route
-                path={routeConfig.WELCOMING_PAGE_ROUTE}
-                render={() => <WelcomingPage søknadstype={søknadstype} onValidSubmit={startSoknad} />}
-            />
+        <Routes>
+            <Route index={true} element={<Navigate to="velkommen" replace={false} />} />
+            <Route path="velkommen" element={<WelcomingPage søknadstype={søknadstype} onValidSubmit={startSoknad} />} />
 
-            {(søknadstype === ApplicationType.pleiepengerBarn ||
-                søknadstype === ApplicationType.pleiepengerLivetsSluttfase) &&
-                isAvailable(søknadstype, StepID.BESKRIVELSE, values) && (
-                    <Route
-                        path={getApplicationRoute(søknadstype, StepID.BESKRIVELSE)}
-                        render={() => (
-                            <BeskrivelseStep
-                                søknadstype={søknadstype}
-                                onValidSubmit={() => navigateToNextStep(StepID.BESKRIVELSE)}
-                            />
-                        )}
-                    />
-                )}
-            {søknadstype === ApplicationType.omsorgspenger && isAvailable(søknadstype, StepID.OMS_TYPE, values) && (
+            {isStepAvailable(søknadstype, StepID.BESKRIVELSE, values) && (
                 <Route
-                    path={getApplicationRoute(søknadstype, StepID.OMS_TYPE)}
-                    render={() => (
+                    path={StepID.BESKRIVELSE}
+                    element={
+                        <BeskrivelseStep
+                            søknadstype={søknadstype}
+                            onValidSubmit={() => navigateToNextStep(StepID.BESKRIVELSE)}
+                        />
+                    }
+                />
+            )}
+            {isStepAvailable(søknadstype, StepID.OMS_TYPE, values) && (
+                <Route
+                    path={StepID.OMS_TYPE}
+                    element={
                         <ValgOmsTypeStep
                             søknadstype={søknadstype}
                             onValidSubmit={() => navigateToNextStep(StepID.OMS_TYPE)}
                         />
-                    )}
+                    }
                 />
             )}
 
-            {isAvailable(søknadstype, StepID.DOKUMENTER, values) && (
+            {isStepAvailable(søknadstype, StepID.DOKUMENTER, values) && (
                 <Route
-                    path={getApplicationRoute(søknadstype, StepID.DOKUMENTER)}
-                    render={() => (
+                    path={StepID.DOKUMENTER}
+                    element={
                         <DokumenterStep
                             søknadstype={søknadstype}
                             onValidSubmit={() => navigateToNextStep(StepID.DOKUMENTER)}
                         />
-                    )}
+                    }
                 />
             )}
 
-            {isAvailable(søknadstype, StepID.OPPSUMMERING, values) && (
+            {isStepAvailable(søknadstype, StepID.OPPSUMMERING, values) && (
                 <Route
-                    path={getApplicationRoute(søknadstype, StepID.OPPSUMMERING)}
-                    render={() => (
+                    path={StepID.OPPSUMMERING}
+                    element={
                         <OppsummeringStep
                             søknadstype={søknadstype}
-                            onApplicationSent={() => {
-                                window.location.href = getRouteUrl(getRouteConfig(søknadstype).APPLICATION_SENDT_ROUTE); // Ensures history is lost
-                            }}
+                            onApplicationSent={() => relocatoToKvitteringPage(søknadstype)}
                         />
-                    )}
+                    }
                 />
             )}
-
-            <Route
-                path={routeConfig.APPLICATION_SENDT_ROUTE}
-                render={() => <ConfirmationPage søknadstype={søknadstype} />}
-            />
-
-            <Route path={routeConfig.ERROR_PAGE_ROUTE} component={GeneralErrorPage} />
-
-            <Redirect to={routeConfig.WELCOMING_PAGE_ROUTE} />
-        </Switch>
+            <Route path={APPLICATION_SENDT_PAGE} element={<ConfirmationPage søknadstype={søknadstype} />} />
+            <Route path={ERROR_PAGE} element={<GeneralErrorPage />} />
+            <Route path="*" element={<Navigate to="" />} />
+        </Routes>
     );
 };