+++ apps/sif-ettersending/src/app/components/pages/confirmation-page/ConfirmationPage.tsx
@@ -26,7 +26,10 @@ const ConfirmationPage = ({ søknadstype }: Props) => {
     return (
         <Page title={intlHelper(intl, 'page.confirmation.sidetittel')} className={bem.block}>
             <div className={bem.element('centeredContent')}>
-                <CheckmarkIcon />
+                <div role="presentation" aria-hidden="true">
+                    <CheckmarkIcon />
+                </div>
+
                 <Block margin="xl">
                     <Heading level="1" size="large">
                         <FormattedMessage id="page.confirmation.tittel" />
@@ -66,7 +69,7 @@ const ConfirmationPage = ({ søknadstype }: Props) => {
                                         target="_blank">
                                         <FormattedMessage id="page.confirmation.check.3.2" />
                                     </Link>
-                                    .
+                                    <FormattedMessage id="page.confirmation.check.3.3" />
                                 </li>
                             </ul>
                         </>