+++ apps/sif-ettersending/src/app/components/step/Step.tsx
@@ -1,20 +1,21 @@
 import { Heading } from '@navikt/ds-react';
-import * as React from 'react';
+import React, { useEffect } from 'react';
 import { useIntl } from 'react-intl';
+import { useNavigate } from 'react-router-dom';
 import BackLink from '@navikt/sif-common-core-ds/lib/components/back-link/BackLink';
 import Page from '@navikt/sif-common-core-ds/lib/components/page/Page';
 import SoknadHeader from '@navikt/sif-common-core-ds/lib/components/soknad-header/SoknadHeader';
 import intlHelper from '@navikt/sif-common-core-ds/lib/utils/intlUtils';
-import { FormikValidationErrorSummary } from '@navikt/sif-common-formik-ds/lib';
-import { History } from 'history';
 import { StepConfigInterface, StepConfigItemTexts, StepID } from '../../config/stepConfig';
 import { getStepTexts } from '../../utils/stepUtils';
 import StepIndicator from './step-indicator/StepIndicator';
+import AriaStepInfo from './aria-step-info/AriaStepInfo';
 
 export interface StepProps {
     id: StepID;
-    useValidationErrorSummary?: boolean;
     bannerTitle?: string;
+    validationSummary?: React.ReactNode;
+    renderAriaStepInfo?: boolean;
 }
 
 interface OwnProps {
@@ -24,26 +25,34 @@ interface OwnProps {
 
 type Props = OwnProps & StepProps;
 
-const Step = ({ id, bannerTitle, stepConfig, useValidationErrorSummary, children }: Props) => {
+const Step = ({ id, bannerTitle, stepConfig, validationSummary, renderAriaStepInfo, children }: Props) => {
     const intl = useIntl();
     const conf = stepConfig[id];
+    const navigate = useNavigate();
     const stepTexts: StepConfigItemTexts = getStepTexts(intl, id, stepConfig);
+
+    useEffect(() => {
+        const stepTitle = document.getElementById('stepTitle');
+        if (stepTitle) {
+            stepTitle.focus();
+        }
+    }, []);
     return (
         <Page
             title={stepTexts.pageTitle}
             topContentRenderer={() => (
                 <>
                     <SoknadHeader title={bannerTitle || intlHelper(intl, 'banner.title')} level="2" />
-                    {useValidationErrorSummary !== false && <FormikValidationErrorSummary />}
+                    {validationSummary !== undefined && validationSummary}
                 </>
             )}>
             {conf.backLinkHref && (
                 <BackLink
                     href={conf.backLinkHref}
                     className="absolute"
-                    onClick={(nextHref: string, history: History, event: React.SyntheticEvent) => {
+                    onClick={(nextHref: string, event: React.SyntheticEvent) => {
                         event.preventDefault();
-                        history.push(nextHref);
+                        navigate(nextHref);
                     }}
                 />
             )}
@@ -51,8 +60,17 @@ const Step = ({ id, bannerTitle, stepConfig, useValidationErrorSummary, children
                 <StepIndicator stepConfig={stepConfig} activeStep={conf.index} />
             </div>
             <div className="mt-12">
-                <Heading level="1" size="large" className="text-center">
+                <Heading
+                    level="1"
+                    size="large"
+                    className="text-center"
+                    style={{ outline: 'none' }}
+                    id="stepTitle"
+                    tabIndex={-1}>
                     {stepTexts.stepTitle}
+                    {renderAriaStepInfo ? (
+                        <AriaStepInfo steps={Object.keys(stepConfig).length} currentStep={conf.index + 1} />
+                    ) : undefined}
                 </Heading>
             </div>
             <div className="mt-8">{children}</div>