+++ apps/sif-ettersending/src/app/config/axiosConfig.ts
@@ -1,4 +1,4 @@
 export default {
-    headers: { 'Content-type': 'application/json; charset=utf-8' },
     withCredentials: true,
+    headers: { 'Content-type': 'application/json; charset=utf-8' },
 };