+++ apps/sif-ettersending/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @navikt/sif-ettersending
 
+## 1.6.0
+
+### Minor Changes
+
+-   509383d: Pakkeoppdateringer, inkl axios
+
+### Patch Changes
+
+-   Updated dependencies [509383d]
+    -   @navikt/sif-common-amplitude@2.6.0
+    -   @navikt/sif-common-core-ds@5.0.0
+    -   @navikt/sif-common-formik-ds@0.8.0
+    -   @navikt/sif-common-sentry@0.5.0
+    -   @navikt/sif-common-utils@3.16.0
+
+## 1.5.0
+
+### Minor Changes
+
+-   7f95812: Router
+
+## 1.4.3
+
+### Patch Changes
+
+-   8533e79: Pakkeoppdateringer
+-   Updated dependencies [8533e79]
+    -   @navikt/sif-common-core-ds@4.4.1
+    -   @navikt/sif-common-formik-ds@0.6.4
+    -   @navikt/sif-common-sentry@0.4.2
+    -   @navikt/sif-common-utils@3.15.3
+
+## 1.4.2
+
+### Patch Changes
+
+-   03aeac4: Oppdatere pakker. Ink. major update av react-router-dom.
+-   Updated dependencies [03aeac4]
+    -   @navikt/sif-common-core-ds@4.4.0
+    -   @navikt/sif-common-amplitude@2.5.2
+    -   @navikt/sif-common-formik-ds@0.6.3
+    -   @navikt/sif-common-sentry@0.4.1
+    -   @navikt/sif-common-utils@3.15.2
+
 ## 1.4.1
 
 ### Patch Changes